+++ programs/pluto/Makefile.options
@@ -320,9 +320,9 @@ DEFINES+=${NSS_FLAGS}    ${FIPS_FLAGS}
 PLUTOMINUSL+=${NSS_LIBS} ${FIPS_LIBS}
 HDRDIRS+=${NSS_HDRDIRS}  ${FIPS_HDRDIRS}
 ifeq ($(USE_NSS),true)
-CRYPTO_OBJS=ikev1_nss.c
+CRYPTO_OBJS=ikev1_nss.c ikev2_nss.c
 else
-CRYPTO_OBJS=ikev1_crypto.o
+CRYPTO_OBJS=ikev1_crypto.o ikev2_crypto.o
 endif
