+++ programs/pluto/ikev2_rsa.c
@@ -60,17 +60,11 @@
 
 #include "oswcrypto.h"
 
-static u_char der_digestinfo[]={
-    0x30, 0x21, 0x30, 0x09, 0x06, 0x05, 0x2b, 0x0e,
-    0x03, 0x02, 0x1a, 0x05, 0x00, 0x04, 0x14
-};
-static int der_digestinfo_len=sizeof(der_digestinfo);
-
-static void ikev2_calculate_sighash(struct state *st
-				    , enum phase1_role role
-				    , unsigned char *idhash
-				    , chunk_t firstpacket
-				    , unsigned char *sig_octets)
+void ikev2_calculate_sighash(struct state *st
+                             , enum phase1_role role
+                             , unsigned char *idhash
+                             , chunk_t firstpacket
+                             , unsigned char *sig_octets)
 {
 	SHA1_CTX       ctx_sha1;
 	const chunk_t *nonce;
@@ -103,129 +97,6 @@ static void ikev2_calculate_sighash(struct state *st
 	SHA1Final(sig_octets, &ctx_sha1);
 }
 
-bool ikev2_calculate_rsa_sha1(struct state *st
-			      , enum phase1_role role
-			      , unsigned char *idhash
-			      , pb_stream *a_pbs)
-{
-	unsigned char  signed_octets[SHA1_DIGEST_SIZE+16];
-	size_t         signed_len;
-	const struct connection *c = st->st_connection;
-	const struct RSA_private_key *k = get_RSA_private_key(c);
-	unsigned int sz;
-
-	if (k == NULL)
-	    return 0;	/* failure: no key to use */
-
-	sz = k->pub.k;
-
-	memcpy(signed_octets, der_digestinfo, der_digestinfo_len);
-
-	ikev2_calculate_sighash(st, role, idhash
-				, st->st_firstpacket_me
-				, signed_octets+der_digestinfo_len);
-	signed_len = der_digestinfo_len + SHA1_DIGEST_SIZE;
-
-	passert(RSA_MIN_OCTETS <= sz && 4 + signed_len < sz && sz <= RSA_MAX_OCTETS);
-
-	DBG(DBG_CRYPT
-	    , DBG_dump("v2rsa octets", signed_octets, signed_len));
-
-	{
-		u_char sig_val[RSA_MAX_OCTETS];
-
-		/* now generate signature blob */
-		sign_hash(k, signed_octets, signed_len
-			  , sig_val, sz);
-		out_raw(sig_val, sz, a_pbs, "rsa signature");
-	}
-
-	return TRUE;
-}
-
-static err_t
-try_RSA_signature_v2(const u_char hash_val[MAX_DIGEST_LEN]
-		     , size_t hash_len
-		     , const pb_stream *sig_pbs, struct pubkey *kr
-		     , struct state *st)
-{
-    const u_char *sig_val = sig_pbs->cur;
-    size_t sig_len = pbs_left(sig_pbs);
-#ifndef HAVE_LIBNSS
-    u_char s[RSA_MAX_OCTETS];	/* for decrypted sig_val */
-    u_char *sig;
-    unsigned int padlen;
-#endif
-    const struct RSA_public_key *k = &kr->u.rsa;
-
-    if (k == NULL)
-	return "1""no key available";	/* failure: no key to use */
-
-    /* decrypt the signature -- reversing RSA_sign_hash */
-    if (sig_len != k->k)
-    {
-	return "1""SIG length does not match public key length";
-    }
-
-#ifdef HAVE_LIBNSS
-    err_t ugh = RSA_signature_verify_nss (k,hash_val,hash_len,sig_val,sig_len);
-    if(ugh!=NULL) {
-	return ugh;
-    }
-#else
-    /* actual exponentiation; see PKCS#1 v2.0 5.1 */
-    {
-	chunk_t temp_s;
-	MP_INT c;
-
-	n_to_mpz(&c, sig_val, sig_len);
-	oswcrypto.mod_exp(&c, &c, &k->e, &k->n);
-
-	temp_s = mpz_to_n(&c, sig_len);	/* back to octets */
-	memcpy(s, temp_s.ptr, sig_len);
-	pfree(temp_s.ptr);
-	mpz_clear(&c);
-    }
-
-    /* check signature contents */
-    /* verify padding */
-    padlen = sig_len - 3 - (hash_len+der_digestinfo_len);
-    /* now check padding */
-    sig = s;
-
-    if(sig[0]    != 0x00
-       || sig[1] != 0x01
-       || sig[padlen+2] != 0x00) {
-	return "2""SIG padding does not check out";
-    }
-
-    /* skip padding */
-    sig += padlen+3;
-
-    /* 2 verify that the has was done with SHA1 */
-    if(memcmp(der_digestinfo, sig, der_digestinfo_len)!=0) {
-	return "SIG not performed with SHA1";
-    }
-
-    sig += der_digestinfo_len;
-
-    DBG(DBG_CRYPT,
-	DBG_dump("v2rsa decrypted SIG:", hash_val, hash_len);
-	DBG_dump("v2rsa computed hash:", sig, hash_len);
-    );
-
-    if(memcmp(sig, hash_val, hash_len) != 0) {
-	return "9""authentication failure: received SIG does not match computed HASH, but message is well-formed";
-    }
-#endif
-
-    unreference_key(&st->st_peer_pubkey);
-    st->st_peer_pubkey = reference_key(kr);
-
-    return NULL;
-}
-
-
 stf_status
 ikev2_verify_rsa_sha1(struct state *st
 		      , enum phase1_role role
@@ -241,6 +112,11 @@ ikev2_verify_rsa_sha1(struct state *st
     invertrole = (role == INITIATOR ? RESPONDER : INITIATOR);
 
     ikev2_calculate_sighash(st, invertrole, idhash, st->st_firstpacket_him, calc_hash);
+
+    DBG(DBG_CRYPT,
+        DBG_dump("v2rsa calculated octets", calc_hash, hash_len);
+        DBG_dump_pbs(sig_pbs);
+        );
 
     return RSA_check_signature_gen(st, calc_hash, hash_len
 				   , sig_pbs
